+++ src/apsbits/core/bluesky_init.py
@@ -1,13 +1,14 @@
 """
-Setup and initialize the Bluesky RunEngine.
-===========================================
-
-This module provides the function init_RE to create and configure a
-Bluesky RunEngine with metadata storage, subscriptions, and various
-settings based on a configuration dictionary.
+Initialize Bluesky functionality for your instrument this includes: Databroker catalog,
+BestEffortCallback: simple real-time visualizations, and Setup and initialization of the
+Bluesky RunEngine.
+==================
 
 .. autosummary::
-    init_RE
+    ~init_catalog
+    ~init_bec_peaks
+    ~init_RE
+
 """
 
 import logging
@@ -17,17 +18,77 @@ from typing import Optional
 from typing import Tuple
 
 import bluesky
+import databroker
+from bluesky.callbacks.best_effort import BestEffortCallback
 from bluesky.utils import ProgressBarManager
 
 from apsbits.utils.controls_setup import connect_scan_id_pv
 from apsbits.utils.controls_setup import set_control_layer
 from apsbits.utils.controls_setup import set_timeouts
+from apsbits.utils.helper_functions import running_in_queueserver
 from apsbits.utils.metadata import get_md_path
 from apsbits.utils.metadata import re_metadata
 from apsbits.utils.stored_dict import StoredDict
 
 logger = logging.getLogger(__name__)
 logger.bsdev(__file__)
+
+TEMPORARY_CATALOG_NAME = "temporalcat"
+
+
+def init_bec_peaks(iconfig):
+    """
+    Create and configure a BestEffortCallback object based on the provided iconfig.
+
+    Parameters:
+        iconfig (dict): Configuration dictionary.
+
+    Returns:
+        tuple: A tuple containing the configured BestEffortCallback object (bec)
+               and its peaks dictionary.
+    """
+
+    bec = BestEffortCallback()
+    """BestEffortCallback object, creates live tables and plots."""
+
+    bec_config = iconfig.get("BEC", {})
+
+    if not bec_config.get("BASELINE", True):
+        bec.disable_baseline()
+
+    if not bec_config.get("HEADING", True):
+        bec.disable_heading()
+
+    if not bec_config.get("PLOTS", True) or running_in_queueserver():
+        bec.disable_plots()
+
+    if not bec_config.get("TABLE", True):
+        bec.disable_table()
+
+    peaks = bec.peaks
+    """Dictionary with statistical analysis of LivePlots."""
+
+    return bec, peaks
+
+
+def init_catalog(iconfig):
+    """
+    Initialize the Databroker catalog using the provided iconfig.
+
+    Parameters:
+        iconfig: Configuration object to retrieve catalog name.
+
+    Returns:
+        Databroker catalog object.
+    """
+    catalog_name = iconfig.get("DATABROKER_CATALOG", TEMPORARY_CATALOG_NAME)
+    try:
+        _cat = databroker.catalog[catalog_name].v2
+    except KeyError:
+        _cat = databroker.temp().v2
+
+    logger.info("Databroker catalog name: %s", _cat.name)
+    return _cat
 
 
 def init_RE(
